Table operations implemented in c++

Table operations implemented in c++. Corresponding R functions without the trailing underscore exist.

tab_perm_(tab, perm)

tab_expand_(tab, aux, type = 0L)

tab_align_(tab1, tab2)

tab_marg_(tab, marg)

tab_op_(tab1, tab2, op = "*")

tab_add_(tab1, tab2)

tab_subt_(tab1, tab2)

tab_mult_(tab1, tab2)

tab_div_(tab1, tab2)

tab_div0_(tab1, tab2)

tab_equal_(tab1, tab2, eps = 1e-12)

tab_list_mult_(lst)

tab_list_add_(lst)

Arguments

tab, tab1, tab2

Tables (arrays)

perm, marg

A vector of indices or dimnames or a right hand sided formula giving the desired permutation/margin.

aux

Either a list with names and dimnames or a named array from which such a list can be extracted.

type

If 0 then entries are duplicated. If 3 then averages are computed. If 2 then 0 slices are inserted.

op

The operation to be carried out; "+", "-", "*", "/".

eps

Criterion for checking equality of two arrays.

lst

List of arrays.