Body Fat Data

Estimates of the percentage of body fat determined by underwater weighing and various body circumference measurements for 252 men.

data(BodyFat)

data(BodyFat)

Format

A data frame with 252 observations on the following 15 variables.

Density

Density determined from underwater weighing, a numeric vector

BodyFat

Percent body fat from Siri's (1956) equation, a numeric vector

Age

in years, a numeric vector

Weight

in lbs, a numeric vector

Height

in inches, a numeric vector

Neck

circumference in cm, a numeric vector

Chest

circumference in cm, a numeric vector

Abdomen

circumference in cm, a numeric vector

Hip

circumference in cm, a numeric vector

Thigh

circumference in cm, a numeric vector

Knee

circumference in cm, a numeric vector

Ankle

circumference in cm, a numeric vector

Biceps

circumference in cm, a numeric vector

Forearm

circumference in cm, a numeric vector

Wrist

circumference in cm, a numeric vector

Source

For more information see https://lib.stat.cmu.edu/datasets/bodyfat

References

Bailey, Covert (1994). Smart Exercise: Burning Fat, Getting Fit, Houghton-Mifflin Co., Boston, pp. 179-186.

Behnke, A.R. and Wilmore, J.H. (1974). Evaluation and Regulation of Body Build and Composition, Prentice-Hall, Englewood Cliffs, N.J.

Siri, W.E. (1956), "Gross composition of the body", in Advances in Biological and Medical Physics, vol. IV, edited by J.H. Lawrence and C.A. Tobias, Academic Press, Inc., New York.

Katch, Frank and McArdle, William (1977). Nutrition, Weight Control, and Exercise, Houghton Mifflin Co., Boston.

Wilmore, Jack (1976). Athletic Training and Physical Fitness: Physiological Principles of the Conditioning Process, Allyn and Bacon, Inc., Boston.

Examples

data(BodyFat)
head(BodyFat)
#>   Density BodyFat Age Weight Height Neck Chest Abdomen   Hip Thigh Knee Ankle
#> 1  1.0708    12.3  23 154.25  67.75 36.2  93.1    85.2  94.5  59.0 37.3  21.9
#> 2  1.0853     6.1  22 173.25  72.25 38.5  93.6    83.0  98.7  58.7 37.3  23.4
#> 3  1.0414    25.3  22 154.00  66.25 34.0  95.8    87.9  99.2  59.6 38.9  24.0
#> 4  1.0751    10.4  26 184.75  72.25 37.4 101.8    86.4 101.2  60.1 37.3  22.8
#> 5  1.0340    28.7  24 184.25  71.25 34.4  97.3   100.0 101.9  63.2 42.2  24.0
#> 6  1.0502    20.9  24 210.25  74.75 39.0 104.5    94.4 107.8  66.0 42.0  25.6
#>   Biceps Forearm Wrist
#> 1   32.0    27.4  17.1
#> 2   30.5    28.9  18.2
#> 3   28.8    25.2  16.6
#> 4   32.4    29.4  18.2
#> 5   32.2    27.7  17.7
#> 6   35.7    30.6  18.8