Personality traits

The peronality dataframe has 240 rows and 32 columns

data(personality)

Format

This dataframe has recordings on the following 32 variables: distant, talkatv, carelss, hardwrk, anxious, agreebl, tense, kind, opposng, relaxed, disorgn, outgoin, approvn, shy, discipl, harsh, persevr, friendl, worryin, respnsi, contrar, sociabl, lazy, coopera, quiet, organiz, criticl, lax, laidbck, withdrw, givinup, easygon

References

Origin unclear

Author

Søren Højsgaard, sorenh@math.aau.dk

Examples

data(personality)
str(personality)
#> 'data.frame':	240 obs. of  32 variables:
#>  $ distant: int  2 3 6 3 7 7 3 7 2 4 ...
#>  $ talkatv: int  7 8 6 7 3 6 6 6 6 7 ...
#>  $ carelss: int  1 2 2 6 3 7 8 2 2 3 ...
#>  $ hardwrk: int  4 7 5 7 5 6 6 9 8 6 ...
#>  $ anxious: int  7 5 1 8 8 7 6 7 2 3 ...
#>  $ agreebl: int  8 8 8 8 6 8 7 8 7 7 ...
#>  $ tense  : int  5 4 2 2 7 7 3 6 4 3 ...
#>  $ kind   : int  9 8 9 8 2 8 5 7 7 7 ...
#>  $ opposng: int  5 5 2 3 3 5 3 2 4 5 ...
#>  $ relaxed: int  6 7 8 7 3 5 7 4 8 6 ...
#>  $ disorgn: int  3 5 7 2 5 6 7 4 3 3 ...
#>  $ outgoin: int  2 8 6 5 2 5 8 6 8 5 ...
#>  $ approvn: int  7 7 7 6 5 6 7 5 7 7 ...
#>  $ shy    : int  9 6 5 4 8 8 5 4 3 5 ...
#>  $ discipl: int  5 7 5 6 7 5 7 8 6 7 ...
#>  $ harsh  : int  5 3 2 2 5 3 2 6 4 6 ...
#>  $ persevr: int  8 7 5 8 6 6 4 8 8 7 ...
#>  $ friendl: int  9 7 8 8 2 7 8 6 8 6 ...
#>  $ worryin: int  9 5 3 3 8 7 6 5 2 4 ...
#>  $ respnsi: int  9 7 9 8 7 7 2 8 8 7 ...
#>  $ contrar: int  5 5 2 2 3 5 2 5 3 4 ...
#>  $ sociabl: int  9 8 8 6 2 4 9 7 8 7 ...
#>  $ lazy   : int  7 2 5 3 6 5 6 4 3 3 ...
#>  $ coopera: int  9 7 8 8 6 7 6 8 8 7 ...
#>  $ quiet  : int  9 6 7 4 7 8 4 6 5 5 ...
#>  $ organiz: int  5 6 4 7 5 6 4 6 7 6 ...
#>  $ criticl: int  5 4 5 6 5 6 6 4 7 5 ...
#>  $ lax    : int  5 3 5 6 3 4 9 3 3 3 ...
#>  $ laidbck: int  7 7 7 6 3 4 9 2 6 4 ...
#>  $ withdrw: int  5 2 6 4 7 6 2 5 3 4 ...
#>  $ givinup: int  5 2 2 3 5 4 3 6 1 2 ...
#>  $ easygon: int  8 7 8 7 5 7 8 6 7 5 ...